% Generated by roxygen2 (4.0.2): do not edit by hand
\name{a.budget}
\alias{a.budget}
\title{Calculate annual energy budget, single function}
\usage{
a.budget(biophys.inputs)
}
\arguments{
\item{biophys.inputs}{Object created from running \code{\link[biophys]{biophys.prep}}}

\item{PFT.out}{Hourly potential foraging times. Pass results from \code{\link[biophys]{pft}}}

\item{metabolism.out}{Hourly metabolic measures. Pass results from \code{\link[biophys]{metabolism}}}
}
\value{
A list that containing a raster stack that includes (1) annual energy budget; (2) total annual metabolic rate; (3) annual potential foraging time; (4) total annual consumption, as well as the annual values in vector format
}
\description{
Calculate annual energy budget, single function
}
\details{
This is a global convenience function to execute \code{\link[biophys]{op.temp.body}}, \code{\link[biophys]{pft}}, \code{\link[biophys]{metabolism}}, and \code{\link[biophys]{budget}}.
}
\author{
Bill Peterman <Bill.Peterman@gmail.com>
}

