% Generated by roxygen2 (4.0.2): do not edit by hand
\name{biophys.prep}
\alias{biophys.prep}
\title{Preparation function to compile biophysical model constants and input.}
\usage{
op.body.temp(Tmax,
Tmin,
elev,
month = NULL,
Julian,
hours,
max.active_temp,
min.active_temp,
NoData.value = -9999,
an.height = 0.005,
wind.height = 10,
u = 1.0,
pCp = 1200,
RH = 0.95,
length.m = 0.055,
phi = 0.623,
SM = 1.5708,
sigma = 5.67e-8,
e = 0.93,
es = 0.96,
cp = 29.3,
alphaS = 0.93,
alphaL = 0.96,
Fp = 0,
Fd.c = 0.8,
Fa.c = 0.5,
Fr.c = 0.5,
Fg.c = 0.5)
}
\arguments{
\item{Tmax}{Maximum temperature raster object. Alternatively, a list of full paths to rasters can be supplied (see Details). The order of list elements must be them same for Tmin, month, and Julian.}

\item{Tmin}{Minimum temperature raster object. Alternatively, a list of full paths to rasters can be supplied (see Details). The order of list elements must be them same for Tmax, month, and Julian.}

\item{elev}{Elevation raster object}

\item{month}{A vector of names to assign to the specified raster layer(s). If multiple months are supplied, the order must be them same as Tmax, Tmin, and Julian.}

\item{Julian}{Julian day (1 - 365). If multiple months are supplied, the order of Julian must be them same as Tmax, Tmin, and month.}

\item{hours}{A vector of values. Specify hours to calculate operative body temperature over.}

\item{max.active_temp}{Maximum temprature that animals will be active (Default = 20)}

\item{min.active_temp}{Minimum temperature that animals will be active (Default = 3)}

\item{NoData.value}{Specify NoData value in raster layers (Default = -9999)}

\item{an.height}{Height of animal off the ground (Default = 0.005)}

\item{wind.height}{Height at which wind speed is measured (Default = 10)}

\item{u}{Measured wind speed at wind meter (Default = 1.0)}

\item{pCp}{Specific heat of air at constant pressure (Default = 1200)}

\item{RH}{Relative humidity (Default = 0.95)}

\item{length.m}{Length of animal in meters (Default = 0.055)}

\item{phi}{Latitude in radians (Default = 0.623)}

\item{SM}{Standard meridian (Default = 1.5708)}

\item{sigma}{Stefan-Boltzmann constant (Default = 5.67e-8)}

\item{e}{Gray body emissivity (Default = 0.93)}

\item{es}{Gray body emissivity (Default = 0.96)}

\item{cp}{Specific heat of air (Default = 29.3)}

\item{alphaS}{Absorptivity in solar waveband (Default = 0.93)}

\item{alphaL}{Absorptivity in thermal waveband (Default = 0.96)}

\item{Fp}{ctor: ratio of projected area perpendicular to solar beam. (Default = 0)}

\item{Fd.c}{Fd constant (Default = 0.8)}

\item{Fa.c}{Fa constant (Default = 0.5)}

\item{Fr.c}{Fr constant (Default = 0.5)}

\item{Fg.c}{Fg constant (Default = 0.5)}
}
\value{
This function returns a list object with the necessary inputs for all other biophys functions.
}
\description{
Preparation function to compile biophysical model constants and input.
}
\details{
Full paths to raster files can be obtain using \code{\link[base]{list.files}}.

 For example: Tmax.files <- list.files(path = "C:/RasterFiles/", pattern = "*max", full.names = TRUE)

 Assuming raster files are stored in a folder called 'RasterFiles', this create the full path to files that include 'max' in the filename. It is advised keep all raster files in a folder separate from other files, and to have a common naming scheme for both minimum and maximum surfaces.
}
\examples{
# Create example data
r.max <- raster(ncol=10, nrow=10)

# assign values to cells
r.max[] <- runif(ncell(r.max),10,20)
r.min <- r.max - 5
r.elev <- r.max * 10

# Make vector of hours
night.hours <- c(19:24,1:7)

biophys.input <- biophys.prep(r.max, r.min, r.elev, hour = night.hours, month = c("April"), Julian = c(106))
}
\author{
Bill Peterman <Bill.Peterman@gmail.com>
}

