% Generated by roxygen2 (4.0.2): do not edit by hand
\name{budget}
\alias{budget}
\title{Calculate annual energy budget}
\usage{
budget(biophys.inputs,
PFT.out,
metabolism.out)
}
\arguments{
\item{biophys.inputs}{Object created from running \code{\link[biophys]{biophys.prep}}}

\item{PFT.out}{Hourly potential foraging times. Pass results from \code{\link[biophys]{pft}}}

\item{metabolism.out}{Hourly metabolic measures. Pass results from \code{\link[biophys]{metabolism}}}
}
\value{
A list that containing a raster stack that includes (1) annual energy budget; (2) total annual metabolic rate; (3) annual potential foraging time; (4) total annual consumption, as well as the annual values in vector format
}
\description{
Calculate annual energy budget
}
\details{
This function summarizes hourly measures to calculate average consumption, average metabolic rate, average potential foraging time, and total metabolic expenditure.
}
\examples{
# Create example data
r.max <- raster(ncol=10, nrow=10)

# assign values to cells
r.max[] <- runif(ncell(r.max),10,20)
r.min <- r.max - 5
r.elev <- r.max * 10

# Make vector of hours
night.hours <- c(19:24,1:7)

biophys.input <- biophys.prep(r.max, r.min, r.elev, hour = night.hours, month = c("April"), Julian = c(106))
temp.out <- op.body.temp(biophys.input)
PFT.out <- pft(biophys.inputs, temp.out)
metabolism.out <- metabolism(biophys.inputs, temp.data)

annual.budget <- budget(biophys.inputs, PFT.out, metabolism.out)
plot(annual.budget$Raster.results)
}
\author{
Bill Peterman <Bill.Peterman@gmail.com>
}

