% Generated by roxygen2 (4.0.1): do not edit by hand
\name{metabolism}
\alias{metabolism}
\title{Calculate active and resting metabolic rates, digestive efficiency, and necessary food consumption}
\usage{
metabolism(biophys.inputs,
temp.data)
}
\arguments{
\item{biophys.inputs}{Object created from running \code{\link[biophys]{biophys.prep}}}

\item{temp.data}{Hourly operative body temperature and air temperature. Pass results from \code{\link[biophys]{op.temp.body}}}
}
\value{
This function returns hourly active and resting metabolic rates, digestive efficiency, and food consumption
}
\description{
Calculate active and resting metabolic rates, digestive efficiency, and necessary food consumption
}
\details{
This function is a wrapper to call an internal function (metab.func)
}
\examples{
# Create example data
r.max <- raster(ncol=10, nrow=10)

# assign values to cells
r.max[] <- runif(ncell(r.max),10,20)
r.min <- r.max - 5
r.elev <- r.max * 10

# Make vector of hours
night.hours <- c(19:24,1:7)

biophys.input <- biophys.prep(r.max, r.min, r.elev, hour = night.hours, month = c("April"), Julian = c(106))
temp.data <- op.body.temp(biophys.input)

metab.out <- metabolism(biophys.inputs, temp.data)
}
\author{
Bill Peterman <Bill.Peterman@gmail.com>
}

