% Generated by roxygen2 (4.0.2): do not edit by hand
\name{op.body.temp}
\alias{op.body.temp}
\title{Calculates operative body temperature}
\usage{
op.body.temp(Tmax,
Tmin,
elev,
NoData.value = -9999,
hours,
an.height = 0.005,
wind.height = 10,
u1 = 1.0,
pCp = 1200,
J = 15,
RH = 0.95,
d1 = 0.055,
phi = 0.623,
SM = 1.5708,
sigma = 5.67e-8,
e = 0.93,
es = 0.96,
cp = 29.3,
alphaS = 0.93,
alphaL = 0.96,
Fp = 0,
Fd.c = 0.8,
Fa.c = 0.5,
Fr.c = 0.5,
Fg.c = 0.5)
}
\arguments{
\item{Tmax}{Maximum temperature raster object}

\item{Tmin}{Minimum temperature raster object}

\item{elev}{Elevation raster object}

\item{NoData.value}{Specify NoData value in raster layers (Default = -9999)}

\item{hours}{A vector of values. Specify hours to calculate operative body temperature over.}

\item{an.height}{Height of animal off the ground (Default = 0.005)}

\item{wind.height}{Height at which wind speed is measured (Default = 10)}

\item{u1}{Measured wind speed at wind meter (Default = 1.0)}

\item{pCp}{Specific heat of air at constant pressure (Default = 1200)}

\item{J}{Julian day of month (Default = 15)}

\item{RH}{Relative humidity (Default = 0.95)}

\item{d1}{Length of animal in meters (Default = 0.055)}

\item{phi}{Latitude in radians (Default = 0.623)}

\item{SM}{Standard meridian (Default = 1.5708)}

\item{sigma}{Stefan-Boltzmann constant (Default = 5.67e-8)}

\item{e}{Gray body emissivity (Default = 0.93)}

\item{es}{Gray body emissivity (Default = 0.96)}

\item{cp}{Specific heat of air (Default = 29.3)}

\item{alphaS}{Absorptivity in solar waveband (Default = 0.93)}

\item{alphaL}{Absorptivity in thermal waveband (Default = 0.96)}

\item{Fp}{ctor: ratio of projected area perpendicular to solar beam. (Default = 0)}

\item{Fd.c}{Fd constant (Default = 0.8)}

\item{Fa.c}{Fa constant (Default = 0.5)}

\item{Fr.c}{Fr constant (Default = 0.5)}

\item{Fg.c}{Fg constant (Default = 0.5)}
}
\value{
This function will return a raster object of operative body temperature, averaged over the specified hours.
}
\description{
Calculates operative body temperature
}
\examples{
# Create example data
r.max <- raster(ncol=10, nrow=10)

# assign values to cells
r.max[] <- runif(ncell(r.max),10,20)
r.min <- r.max - 5
r.elev <- r.max * 10

# Make vector of hours
night.hours <- c(19:24,1:7)

op.temp <- op.body.temp(r.max, r.min, r.elev, hour = night.hours)
plot(op.temp)
}
\author{
Bill Peterman <Bill.Peterman@gmail.com>
}

