% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pft}
\alias{pft}
\title{Calculate Potential Foraging Time (PFT) based on evaporative water loss}
\usage{
pft(biophys.inputs,
temp.data)
}
\arguments{
\item{biophys.inputs}{Object created from running \code{\link[biophys]{biophys.prep}}}

\item{temp.data}{Hourly operative body temperature and air temperature. Pass results from \code{\link[biophys]{op.temp.body}}}
}
\value{
This function will return the hourly evaporative water loss
}
\description{
Calculate Potential Foraging Time (PFT) based on evaporative water loss
}
\details{
This function is a wrapper to call an internal function (evap.func)
}
\examples{
# Create example data
r.max <- raster(ncol=10, nrow=10)

# assign values to cells
r.max[] <- runif(ncell(r.max),10,20)
r.min <- r.max - 5
r.elev <- r.max * 10

# Make vector of hours
night.hours <- c(19:24,1:7)

biophys.input <- biophys.prep(r.max, r.min, r.elev, hour = night.hours, month = c("April"), Julian = c(106))
op.temp <- op.body.temp(biophys.input)

PFT <- pft(biophys.inputs, op.temp)
}
\author{
Bill Peterman <Bill.Peterman@gmail.com>
}

