% Generated by roxygen2 (4.0.1): do not edit by hand
\name{op.body.temp}
\alias{op.body.temp}
\title{Calculates operative body temperature}
\usage{
op.body.temp(biophys.inputs)
}
\arguments{
\item{biophys.inputs}{Object created from running \code{\link[biophys]{biophys.prep}}}
}
\value{
By default, this function will return a list of hourly operative body temperatures for each raster pixel. Additionally, calculated parameters needed for subsequent functions are also saved. The list object generated by this function is formatted for input into \code{\link[biophys]{active.evap}}
}
\description{
Calculates operative body temperature
}
\examples{
# Create example data
r.max <- raster(ncol=10, nrow=10)

# assign values to cells
r.max[] <- runif(ncell(r.max),10,20)
r.min <- r.max - 5
r.elev <- r.max * 10

# Make vector of hours
night.hours <- c(19:24,1:7)

biophys.input <- biophys.prep(r.max, r.min, r.elev, hour = night.hours, month = c("April"), Julian = c(106))
temp.data <- op.body.temp(biophys.input)
}
\author{
Bill Peterman <Bill.Peterman@gmail.com>
}

